SPL-15 * Moved plugin folder
