+++ src/plugin/load-plugin.ts
@@ -1,4 +1,4 @@
-import { IOServiceInterface } from '../../models/services';
+import { IOServiceInterface } from '../models/services';
 
 export function getPluginList(ioService: IOServiceInterface) {
     // console.log(ioService.listDirectories(ioService.getPluginDirectory(), ''));
